Move submodules to Submodules. Build and link dependencies as libraries.
