+++ .gitmodules
@@ -1,42 +1,42 @@
-[submodule "Dependencies/glm"]
-	path = Dependencies/glm
+[submodule "Submodules/GLM"]
+	path = Submodules/GLM
 	url = https://github.com/g-truc/glm.git
-[submodule "Dependencies/glfw"]
-	path = Dependencies/glfw
-	url = https://github.com/glfw/glfw.git
-[submodule "Dependencies/tinyobjloader"]
-	path = Dependencies/tinyobjloader
-	url = https://github.com/tinyobjloader/tinyobjloader.git
-[submodule "Dependencies/stb"]
-	path = Dependencies/stb
-	url = https://github.com/nothings/stb.git
-[submodule "Dependencies/glad"]
-	path = Dependencies/glad
-	url = https://github.com/Dav1dde/glad.git
-[submodule "Dependencies/imgui"]
-	path = Dependencies/imgui
-	url = https://github.com/ocornut/imgui.git
-[submodule "Dependencies/rapidjson"]
-	path = Dependencies/rapidjson
-	url = https://github.com/Tencent/rapidjson.git
-[submodule "Dependencies/tinygltf"]
-	path = Dependencies/tinygltf
-	url = https://github.com/syoyo/tinygltf.git
-[submodule "Dependencies/stduuid"]
-	path = Dependencies/stduuid
-	url = https://github.com/mariusbancila/stduuid.git
-[submodule "Dependencies/entt"]
-	path = Dependencies/entt
+[submodule "Submodules/ENTT"]
+	path = Submodules/ENTT
 	url = https://github.com/skypjack/entt.git
-[submodule "Dependencies/shaderc"]
-	path = Dependencies/shaderc
-	url = https://github.com/google/shaderc.git
-[submodule "Dependencies/glslang"]
-	path = Dependencies/glslang
+[submodule "Submodules/GLFW"]
+	path = Submodules/GLFW
+	url = https://github.com/glfw/glfw.git
+[submodule "Submodules/GLAD"]
+	path = Submodules/GLAD
+	url = https://github.com/Dav1dde/glad.git
+[submodule "Submodules/IMGUI"]
+	path = Submodules/IMGUI
+	url = https://github.com/ocornut/imgui.git
+[submodule "Submodules/RapidJSON"]
+	path = Submodules/RapidJSON
+	url = https://github.com/Tencent/rapidjson.git
+[submodule "Submodules/STB"]
+	path = Submodules/STB
+	url = https://github.com/nothings/stb.git
+[submodule "Submodules/STDUUID"]
+	path = Submodules/STDUUID
+	url = https://github.com/mariusbancila/stduuid.git
+[submodule "Submodules/TinyGLTF"]
+	path = Submodules/TinyGLTF
+	url = https://github.com/syoyo/tinygltf.git
+[submodule "Submodules/TinyOBJLoader"]
+	path = Submodules/TinyOBJLoader
+	url = https://github.com/tinyobjloader/tinyobjloader.git
+[submodule "Submodules/GLSLLang"]
+	path = Submodules/GLSLLang
 	url = https://github.com/KhronosGroup/glslang.git
-[submodule "Dependencies/SPIRV-Tools"]
-	path = Dependencies/SPIRV-Tools
-	url = https://github.com/KhronosGroup/SPIRV-Tools.git
-[submodule "Dependencies/SPIRV-Headers"]
-	path = Dependencies/SPIRV-Headers
+[submodule "Submodules/SPIRVHeaders"]
+	path = Submodules/SPIRVHeaders
 	url = https://github.com/KhronosGroup/SPIRV-Headers.git
+[submodule "Submodules/SPIRVTools"]
+	path = Submodules/SPIRVTools
+	url = https://github.com/KhronosGroup/SPIRV-Tools.git
+[submodule "Submodules/Shaderc"]
+	path = Submodules/Shaderc
+	url = https://github.com/google/shaderc.git
