+++ CMakeLists.txt
@@ -9,7 +9,8 @@ set_property(GLOBAL PROPERTY CMAKE_EXPORT_COMPILE_COMMANDS ON)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/Binaries)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/Libraries)
 
-set(DEP_DIR "${CMAKE_CURRENT_SOURCE_DIR}/Dependencies")
+set(SUBMODULES_DIR "${CMAKE_CURRENT_SOURCE_DIR}/Submodules")
+set(DEPENDENCIES_DIR "${CMAKE_CURRENT_SOURCE_DIR}/Dependencies")
 
 project(Simpleton CXX)
